ove drag select into DragSelectManager, reorganize managers
