+++ src/managers/CommentViewManager.ts
