+++ src/managers/SelectionManager.ts
