+++ src/managers/StageViewManager.ts
