+++ src/managers/TextViewManager.ts
