+++ src/managers/index.ts
@@ -1,8 +1,7 @@
 export { StageViewManager } from "./StageViewManager";
 export { CommentViewManager } from "./CommentViewManager";
-export { handleDragEnd } from "./dragging";
 export { KeyPressManager } from "./KeyPressManager";
 export { ExportManager } from "./ExportManager";
 export { SelectionManager } from "./SelectionManager";
 export { TextViewManager } from "./TextViewManager";
-export { getElementsWithinBounds } from "./dragSelect";
+export { DragSelectManager } from "./DragSelectManager";
